Calculate coancestries

Given genetic data it provides coancestry for between and within populations

calculate_coancestries(genetic_data_parents, genotyped_parent_populations,
  genetic_data_F1 = NA, population_individual_id = NA,
  column_individual = "id", column_population = "population_id",
  pedigree = NA, all_parents_genotyped = FALSE, verbose = TRUE)

Arguments

genetic_data_parents

A data frame or file name containing genetic data for the parental generation. Must be dosage format.

genotyped_parent_populations

A vector of population IDs for the parental generation.

genetic_data_F1

A data frame or file name containing genetic data for the F1 generation. Must be dosage format.

population_individual_id

(Optional) A data frame mapping individuals to their respective populations. First parent ids, then F1 ids. If not provided, a balanced design is assumed.

column_individual

The name of the column containing individual IDs. Default is "id".

column_population

The name of the column containing population IDs. Default is "population_id".

pedigree

(Optional) A data frame containing pedigree information with at least three columns: individual ID, dam ID, and sire ID.

all_parents_genotyped

Logical; if TRUE, indicates that the genotyped individuals in "genotyped_parent_populations" are the parents of F1 individuals

verbose

Logical; if TRUE (default), prints informational messages during calculation. If FALSE, suppresses output.

Value

A list containing:

M

A kinship-based relatedness matrix adjusted for F1 individuals.

Theta.P

An adjusted population coancestry matrix.

Details

This function calculates kinship and coancestry measures from genetic data or pedigree-based information, adjusting for population structure and F1 generation characteristics.

References

Goudet & Weir (2023)

Author

Isabela do O isabela.doo@unil.ch